requirements.txt
# Core dependencies
requests>=2.31.0
pillow>=10.0.0
numpy>=1.24.0

# Windows control
pywin32>=306 ; platform_system == "Windows"

# ShowUI dependencies (optional)
# Install with: pip install -r requirements-showui.txt
# torch>=2.0.0
# torchvision>=0.15.0
# transformers>=4.36.0
# accelerate>=0.25.0
# gradio>=4.0.0

# ASCII tools
opencv-python-headless>=4.8.0

# Development tools
black>=23.0.0
pytest>=7.4.0
mypy>=1.5.0